+++ Tower-Defense/Player.h
@@ -3,10 +3,8 @@
 class Player
 {
 private:
-
 	int life;
 	int money;
-
 
 public:
 
@@ -18,18 +16,18 @@ public:
 
 	//GETTEREK: lekérdezést biztosítja(ezzel jelenítjük majd meg a HUD-on az életet és a pénzt, constans értékek)
 	int getLife() const;
-	int getMoney() const;
+	double getMoney() const;
 	int *getLifePtr();
 	int *getMoneyPtr();
 
 	//SETTEREK: arra kell, hogy tudjuk módosítani a Life és a Money értékén.
 
 	void setLife(int _life);
-	void setMoney(int _money);
+	void setMoney(double _money);
 
 	// "hozzáad / elvesz metórusok" ha kell
-	int sellTower(int _money);
-	int buyTower(int _money);
+	double sellTower(double _money);
+	double buyTower(double _money);
 
 	//vásárlás metódusa
 	//void pay(int _money);
